requirements.txt
SpeechRecognition
google-generativeai
edge-tts
pygame
pyaudio
pyautogui
pygetwindow
python-dotenv
asyncio
flet
pyinstaller
